eval(): Auto-detect basic expressions.
eval() now uses the parser to determine the type of the input code. For
basic types and variables, it seems to be succeeding.

Support for code blocks is still a work in progress. The detection fails
due to the lack of a real parent node, and incorrect indentation at
times.

Due to detection failures, code blocks are currently evaluated as a
fallback when every other test fails.